suzerain: fix unsigned/signed warning
git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@44839 0bd85f0b-c151-43b9-a1e9-27231104551f

+++ apps/reacting/driver.cpp
@@ -220,7 +220,7 @@ driver::save_spectra_primitive(
     if (Ns > 1) {
         assert(static_cast<int>(ndx::species) < swave_count);
     }
-    SUZERAIN_ENSURE((int) state_nonlinear->shape()[0] == swave_count);
+    SUZERAIN_ENSURE(state_nonlinear->shape()[0] == swave_count);
     SUZERAIN_ENSURE(std::equal(state_nonlinear->shape() + 1, 
                                state_nonlinear->shape() + 4,
                                auxw.shape() + 1));
